[Feature][Maxell] removed google maps
