+++ package.json
@@ -1,61 +1,61 @@
 {
-  "name": "hakbang",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "next dev --turbopack",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "geojson": "node scripts/split-linestring.js",
-    "distance": "node scripts/test-distance.js",
-    "nearest-places": "node scripts/extract-nearest-places.js",
-    "weather": "node scripts/extract-weather.js"
-  },
-  "dependencies": {
-    "@aws-sdk/client-s3": "^3.675.0",
-    "@aws-sdk/s3-request-presigner": "^3.675.0",
-    "@emotion/react": "^11.13.3",
-    "@emotion/styled": "^11.13.0",
-    "@headlessui/react": "^2.1.10",
-    "@iconify/react": "^5.0.2",
-    "@mui/material": "^6.1.4",
-    "@react-google-maps/api": "^2.20.3",
-    "@turf/turf": "^7.1.0",
-    "@types/googlemaps": "^3.43.3",
-    "@vercel/analytics": "^1.4.0",
-    "aws-sdk": "^2.1691.0",
-    "axios": "^1.7.7",
-    "dotenv": "^16.4.5",
-    "driver.js": "^1.3.1",
-    "framer-motion": "^11.11.9",
-    "lodash.debounce": "^4.0.8",
-    "multer": "^1.4.5-lts.1",
-    "multer-s3": "^3.0.1",
-    "next": "15.1.6",
-    "nookies": "^2.5.2",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "sass": "^1.80.6",
-    "zustand": "^5.0.0"
-  },
-  "devDependencies": {
-    "@types/geojson": "^7946.0.14",
-    "@types/google.maps": "^3.58.1",
-    "@types/lodash.debounce": "^4.0.9",
-    "@types/node": "^20",
-    "@types/react": "19.0.8",
-    "@types/react-dom": "19.0.3",
-    "eslint": "^8",
-    "eslint-config-next": "15.1.6",
-    "postcss": "^8",
-    "prettier": "3.3.3",
-    "tailwindcss": "^3.4.1",
-    "typescript": "^5"
-  },
-  "overrides": {
-    "@types/react": "19.0.8",
-    "@types/react-dom": "19.0.3"
-  }
+    "name": "lakbai",
+    "version": "0.1.0",
+    "private": true,
+    "type": "module",
+    "scripts": {
+        "dev": "next dev --turbopack",
+        "build": "next build",
+        "start": "next start",
+        "lint": "next lint",
+        "geojson": "node scripts/split-linestring.js",
+        "distance": "node scripts/test-distance.js",
+        "nearest-places": "node scripts/extract-nearest-places.js",
+        "weather": "node scripts/extract-weather.js"
+    },
+    "dependencies": {
+        "@aws-sdk/client-s3": "^3.675.0",
+        "@aws-sdk/s3-request-presigner": "^3.675.0",
+        "@emotion/react": "^11.13.3",
+        "@emotion/styled": "^11.13.0",
+        "@headlessui/react": "^2.1.10",
+        "@iconify/react": "^5.0.2",
+        "@mui/material": "^6.1.4",
+        "@react-google-maps/api": "^2.20.3",
+        "@turf/turf": "^7.1.0",
+        "@types/googlemaps": "^3.43.3",
+        "@vercel/analytics": "^1.4.0",
+        "aws-sdk": "^2.1691.0",
+        "axios": "^1.7.7",
+        "dotenv": "^16.4.5",
+        "driver.js": "^1.3.1",
+        "framer-motion": "^11.11.9",
+        "lodash.debounce": "^4.0.8",
+        "multer": "^1.4.5-lts.1",
+        "multer-s3": "^3.0.1",
+        "next": "15.1.6",
+        "nookies": "^2.5.2",
+        "react": "19.0.0",
+        "react-dom": "19.0.0",
+        "sass": "^1.80.6",
+        "zustand": "^5.0.0"
+    },
+    "devDependencies": {
+        "@types/geojson": "^7946.0.14",
+        "@types/google.maps": "^3.58.1",
+        "@types/lodash.debounce": "^4.0.9",
+        "@types/node": "^20",
+        "@types/react": "19.0.8",
+        "@types/react-dom": "19.0.3",
+        "eslint": "^8",
+        "eslint-config-next": "15.1.6",
+        "postcss": "^8",
+        "prettier": "3.3.3",
+        "tailwindcss": "^3.4.1",
+        "typescript": "^5"
+    },
+    "overrides": {
+        "@types/react": "19.0.8",
+        "@types/react-dom": "19.0.3"
+    }
 }
